+++ resources/aws_ses_dedicated_ip_pool.rb
@@ -40,5 +40,5 @@ rest_property_map({
 })
 
 rest_post_only_properties %i{
-  pool_name scaling_mode
+  pool_name
 }